requirements.txt
pyrogram
tgcryto
python-dotenv
smsactivateru
pyromod
faker